fix: saving image

+++ pages/admin/hotel.vue
@@ -435,10 +435,8 @@ export default {
           }))
         )
         this.item.imagesMeta = _cloneDeep(imagesMeta)
-        this.itemOriginal.imagesMeta = _cloneDeep(imagesMeta)
       } else {
         this.item.imagesMeta = []
-        this.itemOriginal.imagesMeta = []
       }
     }
   },
@@ -643,11 +641,15 @@ export default {
             createdAt: this.isEditing ? this.item.createdAt : date,
             updatedAt: date
           }
-          if (this.isEdited) {
+          if (
+            this.isEdited &&
+            this.itemOriginal.imagesMeta &&
+            this.itemOriginal.imagesMeta.length > 0
+          ) {
             await Promise.all(
-              this.itemOriginal.imagesMeta.map(meta =>
-                storage.ref(meta.fullPath).delete()
-              )
+              this.itemOriginal.imagesMeta.map(meta => {
+                return storage.ref(meta.fullPath).delete()
+              })
             )
           }
           const imagesMeta = await Promise.all(
@@ -689,6 +691,7 @@ export default {
           await this.$notify({ kind: 'success', message: this.$t('dataSaved') })
         }
       } catch (error) {
+        console.log(error)
         this.$notify({
           isError: true,
           message: error.message
@@ -712,9 +715,11 @@ export default {
           .collection(this.collection)
           .doc(item.uid)
           .delete()
-        await Promise.all(
-          item.imagesMeta.map(meta => storage.ref(meta.fullPath).delete())
-        )
+        if (item.imagesMeta && item.imagesMeta.length > 0) {
+          await Promise.all(
+            item.imagesMeta.map(meta => storage.ref(meta.fullPath).delete())
+          )
+        }
         await this.getItems()
         await this.onDeleteClose()
         await this.$notify({ kind: 'success', message: this.$t('dataDeleted') })
